Fix chan scaled loss calculation

+++ sclassifier/feature_extractor_ae.py
@@ -1095,12 +1095,11 @@ class FeatExtractorAE(object):
 			###tf.print("--> MSE loss: data_abs_max=", data_abs_max, output_stream=sys.stdout)
 			
 			y_true_safe= tf.where(~cond, tf.ones_like(y_true) * 0, y_true)
-			#y_true_nonzeros= tf.cast(y_true_safe!= 0, tf.float32)
-			y_true_nonzeros= tf.cast(y_true_safe!=0, tf.float64)
+			y_true_nonzeros= tf.cast(y_true_safe!= 0, tf.float32)
 			n_nonzeros= tf.reduce_sum(y_true_nonzeros, axis=(1,2))
 			n_sum= tf.reduce_sum(y_true_safe, axis=(1,2)) 
 			#data_means= tf.math.divide_no_nan(n_sum, n_nonzeros) # NB: This returns 0 if n_nonzeros=0
-			data_means= tf.math.divide_no_nan(tf.cast(n_sum, tf.float64), tf.cast(n_nonzeros, tf.float64) ) # NB: This returns 0 if n_nonzeros=0
+			data_means= tf.math.divide_no_nan(tf.cast(n_sum, tf.float32), tf.cast(n_nonzeros, tf.float32) ) # NB: This returns 0 if n_nonzeros=0
 			data_means_max= tf.reduce_max(data_means, axis=1)
 			data_means_max= tf.expand_dims(data_means_max, axis=1)
 			##chan_weights= tf.where(data_means==0, tf.ones_like(data_means)*1, data_means_max/data_means)
